Pair-programmer/Pair_Programmer/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
google-generativeai==0.3.1
python-multipart==0.0.6
httpx==0.25.1
aiohttp==3.9.1 